Teach the script to automatically sync data that is older than what is in the database
This commit teaches the script to download data earlier than what is recorded in the database. Thus every sync attempt will now attempt to download data that is newer *and* older than what is already in the database. This change was necessary to continue previous download attempts that were cut short, for example, by rate limiting.
Fixes #2. Thanks to Amir Tahvildaran for the problem report!